Refactor code and file structure.
