+++ manage.py
@@ -3,11 +3,11 @@ import unittest
 from flask_migrate import Migrate, MigrateCommand
 from flask_script import Manager
 
-from api.main import create_app
-from api.main.extensions import db
+from zimmerman import create_app
+from zimmerman.extensions import db
 
 # Import blueprint
-from api import main_bp
+from zimmerman.api import main_bp
 
 # Create the application in development api.
 # We obviously want to change this to 'prod' in deployment.
@@ -33,7 +33,7 @@ def run():
 @manager.command
 def test():
     """ Runs Unit Tests """
-    tests = unittest.TestLoader().discover("tests", pattern="test*.py")
+    tests = unittest.TestLoader().discover("zimmerman/tests", pattern="test*.py")
     result = unittest.TextTestRunner(verbosity=2).run(tests)
     if result.wasSuccessful():
         return 0
