+++ zimmerman/api/main/controller/like_controller.py
@@ -1,7 +1,7 @@
 from flask_restx import Resource
 from flask_jwt_extended import jwt_required, get_jwt_identity
 
-from api.main import limiter
+from zimmerman import limiter
 
 from ..util.dto import LikeDto
 from ..service.like_service import Like, Unlike
