+++ zimmerman/api/main/controller/upload_controller.py
@@ -2,7 +2,7 @@ from flask import request
 from flask_restx import Resource
 from flask_jwt_extended import jwt_required
 
-from api.main import limiter
+from zimmerman import limiter
 from ..util.dto import UploadDto
 from ..service.upload_service import upload_file
