+++ zimmerman/api/notification/controller.py
@@ -2,8 +2,8 @@ from flask import request
 from flask_restx import Resource
 from flask_jwt_extended import jwt_required, get_jwt_identity
 
-from api.main import limiter
-from api.main.service.user.utils import load_user
+from zimmerman import limiter
+from zimmerman.api.main.service.user.utils import load_user
 
 from .service import NotificationService
 from .util.dto import NotificationDto
