+++ zimmerman/models/notification.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from api.main import db
+from zimmerman import db
 
 # Alias common SQLAlchemy names
 Column = db.Column
