+++ zimmerman/tests/test_config.py
@@ -4,12 +4,12 @@ from flask import current_app
 from flask_testing import TestCase
 
 from manage import app
-from api.config import basedir
+from zimmerman.config import basedir
 
 
 class TestDevelopmentConfig(TestCase):
     def create_app(self):
-        app.config.from_object("api.config.DevelopmentConfig")
+        app.config.from_object("zimmerman.config.DevelopmentConfig")
         return app
 
     def test_app_is_development(self):
@@ -24,7 +24,7 @@ class TestDevelopmentConfig(TestCase):
 
 class TestProductionConfig(TestCase):
     def create_app(self):
-        app.config.from_object("api.config.ProductionConfig")
+        app.config.from_object("zimmerman.config.ProductionConfig")
         return app
 
     def test_app_is_production(self):
