+++ zimmerman/tests/test_like.py
@@ -3,8 +3,8 @@ import json
 from datetime import datetime
 
 from flask import current_app
-from tests.base import BaseTestCase
-from tests.common_functions import register_user, login_user
+from zimmerman.tests.base import BaseTestCase
+from zimmerman.tests.common_functions import register_user, login_user
 
 
 def create_post(self, access_token):
